Late Move Reduction ajustment
In the last major commit, i fixed a bug where the move ordering would return a random negative number a lot. This caused extra pruning, and made a better search. When I fixed this bug, the search got worse, and i just figured out why and fixed it. wow that was frustrating.